SQUASH – move SSTableHeaderFix.java to o.a.c.tools package, as it's only usage now is sstableupgrade
